requirements.txt
streamlit>=1.36
pydantic>=2.7
anthropic>=0.40
weasyprint>=62.0
jinja2>=3.1
google-api-python-client>=2.140
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
python-dotenv>=1.0
pandas>=2.2
